MarcoAmancha

+++ proyecto_integrador/src/proyecto_integrador/dao/IUsuario.java
@@ -7,7 +7,7 @@ public interface IUsuario {
     public int insertar(Usuario usuario) throws Exception;
     public int modificar(Usuario usuario) throws Exception;
     public int eliminar(Usuario usuario) throws Exception;
-    public Usuario obtener(int codigo) throws Exception;
+    public Usuario obtener(long codigo) throws Exception;
     public List<Usuario> obtener() throws Exception; 
     
 }
